requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.6.0
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
decorator==5.1.1
Django==4.2.6
django-celery-beat==2.5.0
django-cors-headers==4.3.0
django-filter==23.3
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
executing==2.0.1
frozenlist==1.4.0
idna==3.4
inflection==0.5.1
ipython==8.17.2
jedi==0.19.1
kombu==5.3.2
load-dotenv==0.1.0
magic-filter==1.0.12
matplotlib-inline==0.1.6
multidict==6.0.4
packaging==23.2
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
prompt-toolkit==3.0.39
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.16.1
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
salary==2.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.13.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.0.0
wcwidth==0.2.8
yarl==1.9.2
